backend/requirements.txt
fastapi
uvicorn
python-multipart
yt-dlp
ffmpeg-python
transformers
torch
torchaudio
openai-whisper
sentence-transformers
faiss-cpu
reportlab
stripe
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
httpx
python-dotenv
